Add classic rules and battle setup flow
- Preserve current combat as default rules
- Add classic armor-based battle system
- Add ruleset selection in UI
- Add Hyoga and rebalance classic roster
- Add play again and main menu

+++ src/main/java/seiya/actions/Gather.java
@@ -3,9 +3,9 @@ package seiya.actions;
 import seiya.characters.Character;
 
 public class Gather extends Action {
-    private final int spiritGain;
+    private final double spiritGain;
 
-    public Gather(int spiritGain) {
+    public Gather(double spiritGain) {
         super("Gather", 0.0, 0.0);
         this.spiritGain = spiritGain;
     }
